warnings removed
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@3056 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/gui/dialogs/GUIDialog_Breakpoints.cpp
@@ -23,6 +23,9 @@ namespace
     "$Id$";
 }
 // $Log$
+// Revision 1.15  2006/11/16 12:30:53  dkrajzew
+// warnings removed
+//
 // Revision 1.14  2006/07/06 05:35:54  dkrajzew
 // replaced exception throwing in unreachable places by something more friendly
 //
@@ -376,7 +379,7 @@ GUIDialog_Breakpoints::onCmdEditTable(FXObject*,FXSelector,void*data)
         value = INVALID_VALUE_STR;
     }
     int row = i->row;
-    if(row==gBreakpoints.size()) {
+    if(row==(int) gBreakpoints.size()) {
         gBreakpoints.push_back(INVALID_VALUE);
     }
